Ench/bill details UI (#103)
* perf: refactor bill details

* feat(complated bill): added created date
