+++ src/components/CancelledBIllTable.js
@@ -24,7 +24,7 @@ const CancelledBillTable = ({ headCells }) => {
           phone: row.patient_phone,
           address: row.patient_address,
           totalAmount: row.total_amount,
-          date: row.created_time.split("T")[0],
+          created_date: row.created_time.split("T")[0],
         };
       });
       setRows(data);
